Move withFileLock to ghc-tags-plugin
