+++ ghc-tags-core/lib/Plugin/GhcTags/Utils.hs
@@ -3,16 +3,7 @@
 module Plugin.GhcTags.Utils
   ( endOfLine
   , notNewLine
-
-  , withFileLock
-  , LockMode (..)
   ) where
-
-import           Control.Exception
-import           System.IO
-import           GHC.IO.Handle
-import           GHC.IO.Handle.Lock
-
 
 -- | Platform dependend eol:
 --
@@ -32,17 +23,3 @@ endOfLine = "\n"
 
 notNewLine :: Char -> Bool
 notNewLine = \x -> x /= '\n' && x /= '\r'
-
-
--- | 'flock' base lock (on posix) or `LockFileEx` on Windows.
---
-withFileLock :: FilePath -> LockMode -> IOMode -> (Handle -> IO x) -> IO x
-withFileLock path mode iomode k =
-    bracket
-      (openFile path iomode)
-      (\h -> hClose h)
-      $ \h ->
-        bracket
-          (hLock h mode)
-          (\_ -> hUnlock h)
-          (\_ -> k h)
